Add tab view to DNS page
